+++ components/http/okHttp/src/main/java/com/microsoft/kiota/http/OkHttpRequestAdapter.java
@@ -61,7 +61,7 @@ import io.opentelemetry.context.Context;
 public class OkHttpRequestAdapter implements com.microsoft.kiota.RequestAdapter {
     private static final String contentTypeHeaderKey = "Content-Type";
     @Nonnull
-    private final OkHttpClient client;
+    private final Call.Factory client;
     @Nonnull
     private final AuthenticationProvider authProvider;
     @Nonnull
@@ -113,7 +113,7 @@ public class OkHttpRequestAdapter implements com.microsoft.kiota.RequestAdapter 
      * @param client the http client to use for sending requests.
      */
     @SuppressWarnings("LambdaLast")
-    public OkHttpRequestAdapter(@Nonnull final AuthenticationProvider authenticationProvider, @Nullable final ParseNodeFactory parseNodeFactory, @Nullable final SerializationWriterFactory serializationWriterFactory, @Nullable final OkHttpClient client) {
+    public OkHttpRequestAdapter(@Nonnull final AuthenticationProvider authenticationProvider, @Nullable final ParseNodeFactory parseNodeFactory, @Nullable final SerializationWriterFactory serializationWriterFactory, @Nullable final Call.Factory client) {
         this(authenticationProvider, parseNodeFactory, serializationWriterFactory, client, null);
     }
     /**
@@ -125,7 +125,7 @@ public class OkHttpRequestAdapter implements com.microsoft.kiota.RequestAdapter 
      * @param observabilityOptions the observability options to use for sending requests.
      */
     @SuppressWarnings("LambdaLast")
-    public OkHttpRequestAdapter(@Nonnull final AuthenticationProvider authenticationProvider, @Nullable final ParseNodeFactory parseNodeFactory, @Nullable final SerializationWriterFactory serializationWriterFactory, @Nullable final OkHttpClient client, @Nullable final ObservabilityOptions observabilityOptions) {
+    public OkHttpRequestAdapter(@Nonnull final AuthenticationProvider authenticationProvider, @Nullable final ParseNodeFactory parseNodeFactory, @Nullable final SerializationWriterFactory serializationWriterFactory, @Nullable final Call.Factory client, @Nullable final ObservabilityOptions observabilityOptions) {
         this.authProvider = Objects.requireNonNull(authenticationProvider, "parameter authenticationProvider cannot be null");
         if(client == null) {
             this.client = KiotaClientFactory.create().build();